+++ http_cache_client/CHANGELOG.md
@@ -3,6 +3,7 @@
 - fix: Improve request date assignment.
 - fix: bump http to ^1.2.0
 - fix: `CacheOptions#maxStale` usage may lead to empty response body.
+- feat: Improve performance on cache check when using cipher.
 
 ## 1.0.0
 - Initial version.