+++ http_cache_core/CHANGELOG.md
@@ -1,6 +1,7 @@
-## 1.0.0
+## 1.0.1
 - chore: Add writeContent to CacheResponse.
 - fix: `CacheOptions#maxStale` usage may lead to empty response body.
+- chore: Allow to read only headers or body.
 
 ## 1.0.0
 - Initial version.